refactor(core): extract fetch error troubleshooting
